examples/beta/cdc-postgres/pyproject.toml
[tool.poetry]
name = "cdc-postgres"
version = "0.1.0"
description = ""
authors = ["greenep12 <ericpgreen2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beneath = "^1.4.2"
psycopg2 = "^2.9.1"
PyYAML = "^5.4.1"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
